Update Fragne Front End Page 1 - Final
Finalized the Log in Page, can add more design elements later like shadowing and stuff like that. The ribbon is basically impossible to recreate with this current state. I can keep trying later on that as well. The font color is still not blue, but it's alright, it still looks good.