Made some privacy adjustments (check description)
announcements section would show the user's email, not their first and last name, we might have to update the backend for this so it can fetch user_id.

In PageUnifiedNetworking it wouldn't show their first and last name so I think this is a backend issue maybe. I updated it to hopefully reflect the users actual first name and last name similar to how the "Connect" and "Mentors" section display names.

PageMessages would also leak the user's email address when you search for their name in the search bar, so I fixed this really quick.